Fix Redox compilation if Redox becomes a part of the unix target family

+++ src/sys/orbital.rs
@@ -136,7 +136,7 @@ impl Window {
             "orbital:{}/{}/{}/{}/{}/{}",
             flag_str, x, y, w, h, title
         )) {
-            if let Ok(address) = unsafe { syscall::fmap(file.as_raw_fd(), &syscall::Map {
+            if let Ok(address) = unsafe { syscall::fmap(file.as_raw_fd() as usize, &syscall::Map {
                 offset: 0,
                 size: (w * h * 4) as usize,
                 flags: syscall::PROT_READ | syscall::PROT_WRITE,
@@ -221,7 +221,7 @@ impl Window {
             .write(&format!("S,{},{}", width, height).as_bytes());
         self.sync_path();
         unsafe {
-            let address = syscall::fmap(self.file.as_raw_fd(), &syscall::Map {
+            let address = syscall::fmap(self.file.as_raw_fd() as usize, &syscall::Map {
                 offset: 0,
                 size: (self.w * self.h * 4) as usize,
                 flags: syscall::PROT_READ | syscall::PROT_WRITE,
